5-经典群体遗传学计算/requirements.txt
numpy==2.1.3
pandas==2.2.3
scikit-allel==1.3.10
msprime==1.3.2
tskit==0.5.8
